requirements.txt
matplotlib==3.6.2
numpy==1.23.5
Pillow==9.3.0
tqdm==4.64.1
wandb==0.13.5
torch==2.7.1
torchvision==0.22.1
transformers==2.1.1 